Extract ModuleMatcher
