+++ B9PartSwitch/B9PartSwitch.csproj
@@ -108,6 +108,7 @@
     <Compile Include="Fishbones\OperationManager.cs" />
     <Compile Include="Fishbones\UseParser.cs" />
     <Compile Include="Localization.cs" />
+    <Compile Include="ModuleMatcher.cs" />
     <Compile Include="PartSwitch\PartModifiers\AttachNodeSizeModifier.cs" />
     <Compile Include="Utils\ChangeTransactionManager.cs" />
     <Compile Include="ModuleB9DisableTransform.cs" />
